refactor: Restructure project for making space
