+++ src/main.rs
@@ -1,6 +1,5 @@
-use zero2prod::run;
-
 use std::net::TcpListener;
+use zero2prod::startup::run;
 
 #[tokio::main]
 async fn main() -> std::io::Result<()> {
